Add confirm dialog
Add a windowClosingListener to the LetUsProunceView, which gives this:
When user is trying to exit the system, a confirm dialog would show up
with a "yes" and a "no" button asking user "Exit system?". And if the
"yes" button is pressed, the system shall close. If the "no" button is
pressed, the system will not exit.